Create api sub folder under src folder and move api module there
